refactor: LottoNumbers 클래스 분리
